# server/openapi.json
{
  "components": {
    "schemas": {
...
      "CompleteRequest": {
        "properties": {
          "signature": {
            "description": "The signature over `to_sign_hash`, base64. PKCS#1 v1.5 for RSA, a DER ECDSA-Sig-Value for EC.",
            "title": "Signature",
            "type": "string"
          }
        },
        "required": [
          "signature"
        ],
        "title": "CompleteRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SignServerSideRequest": {
        "properties": {
          "document": {
            "description": "The document, base64.",
            "title": "Document",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SigningOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "document"
        ],
        "title": "SignServerSideRequest",
        "type": "object"
      },
...
      "StartSignatureRequest": {
        "properties": {
          "certificate": {
            "description": "The signer's certificate, base64 DER.",
            "title": "Certificate",
            "type": "string"
          },
          "document": {
            "description": "The PDF, base64.",
            "title": "Document",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SigningOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "document",
          "certificate"
        ],
        "title": "StartSignatureRequest",
        "type": "object"
      },
      "StoredDocument": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          },
          "download_url": {
            "examples": [
              "/api/documents/abc123"
            ],
            "title": "Download Url",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "download_url"
        ],
        "title": "StoredDocument",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The document never leaves this server. A browser carries a 32-byte hash out and a signature back, so a 200 MB file signs as fast as a 200 KB one.\n\nThe protocol is frozen in CONTRACTS.md section 1. Generate a client from this document rather than hand-writing one.",
    "summary": "Sign PDFs, any file (CAdES) and XML (XAdES) with a token or a server-held key.",
    "title": "signer-server",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/cades/sign-server-side": {
      "post": {
        "operationId": "cades_server_side_api_cades_sign_server_side_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignServerSideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad document, certificate or signature"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such session or document"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session or document expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Produce a .p7s with the server's own key",
        "tags": [
          "CAdES"
        ]
      }
    },
    "/api/cades/signatures": {
      "post": {
        "operationId": "start_cades_api_cades_signatures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSignatureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad document, certificate or signature"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such session or document"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session or document expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Start a detached CAdES session over any file",
        "tags": [
          "CAdES"
        ]
      }
    },
    "/api/cades/signatures/{session_id}/complete": {
      "post": {
        "operationId": "complete_cades_api_cades_signatures__session_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredDocument"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad document, certificate or signature"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such session or document"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session or document expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Finish a CAdES session, producing a .p7s",
        "tags": [
          "CAdES"
        ]
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/pkcs7-signature": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "application/xml": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The signed document"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad document, certificate or signature"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such session or document"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session or document expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Download a signed document",
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/sign-server-side": {
      "post": {
        "operationId": "sign_server_side_api_sign_server_side_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignServerSideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSideSigned"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad document, certificate or signature"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such session or document"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session or document expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Sign a PDF with the server's own key",
        "tags": [
          "PDF"
        ]
      }
    },
    "/api/signatures": {
      "post": {
        "operationId": "start_signature_api_signatures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSignatureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad document, certificate or signature"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such session or document"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Session or document expired"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Start a token signing session",
        "tags": [
          "PDF"
        ]
      }
    },
...
  }
}